feat: add CSPViolationBox
