+++ features/ipfs/csp-violation-box/index.tsx
@@ -1,0 +1,1 @@
+export * from './csp-violation-box';
